Finished routes working with frontend

+++ server/routes/entry.js
@@ -3,11 +3,13 @@ const express = require('express');
 const router = express.Router();
 const Entry = require('../models/entry');
 const catchAsync = require('../utils/catchAsync');
+const convertToUTC = require('../utils/convertToUTC');
 const validateEntry = require('../middleware/validateEntry');
 const { storage } = require('../cloudinary');
 const multer = require('multer');
 const upload = multer({ storage: storage });
 const uuid = require('uuid');
+const { cloudinary } = require('../cloudinary');
 
 router.route('/')
     // Get all entries
@@ -31,7 +33,7 @@ router.route('/')
                 type: reqEntry.type,
                 title: reqEntry.title,
                 artist: reqEntry.artist,
-                releaseDate: new Date(reqEntry.releaseDate),
+                releaseDate: convertToUTC(new Date(reqEntry.releaseDate)),
                 genre: reqEntry.genre,
                 cover: {
                     url: result.path,
@@ -63,17 +65,52 @@ router.route('/:id')
     }))
     // Replace / update an entry
     .put(validateEntry, catchAsync(async (req, res, next) => {
+        // Entry from the request and add releaseDate and reviewDate to prevent day errors
+        reqEntry = JSON.parse(req.body.entry);
+        // If dates aren't in UTC change them
+        const updatedEntry = {
+            id: reqEntry.id,
+            reviewed: reqEntry.reviewed,
+            type: reqEntry.type,
+            title: reqEntry.title,
+            artist: reqEntry.artist,
+            releaseDate: convertToUTC(new Date(reqEntry.releaseDate)),
+            genre: reqEntry.genre,
+            cover: {
+                url: reqEntry.cover.url,
+                filename: reqEntry.cover.filename
+            },
+            review: {
+                rating: reqEntry.review.rating,
+                reviewDate: convertToUTC(new Date(reqEntry.review.reviewDate)),
+                reviewText: reqEntry.review.reviewText
+            }
+        }
+        // console.log("Updated entry: " + updatedEntry);
+        // Get the id and update the entry
         const { id } = req.params;
-        const entry = await Entry.findOneAndUpdate({ id }, { ...req.body.entry }, { new: true });
-        console.log("req.body.entry: " + req.body.entry)
-        console.log("Updating entry with id: " + req.params.id);
+        const entry = await Entry.findOneAndUpdate({ id }, updatedEntry, { new: true });
+        // console.log("Updated entry: " + entry);
+        // console.log("Updating entry with id: " + req.params.id);
         res.json(entry);
     }))
     // Delete an entry
     .delete(catchAsync(async (req, res, next) => {
         const { id } = req.params;
+        // Find the Entry by ID
+        const entry = await Entry.findOne({ id });
+        if (!entry) {
+            return res.status(404).json({ message: 'Entry not found' });
+        }
+        // Delete the image from Cloudinary using the filename
+        const filename = entry.cover.filename;
+        if (filename) {
+            // Delete the image from Cloudinary
+            await cloudinary.uploader.destroy(filename);
+        }
+        // Delete the entry from the database
         await Entry.findOneAndDelete({ id });
-        console.log("Deleting entry with id: " + req.params.id);
+        // console.log("Deleting entry with id: " + req.params.id);
         res.json({ message: 'Deleted' });
     }));
